rewrite SelectPattern and CanApply to cut down on iterations
there is now a bug where the candidatePosition is a random position not adjacent to any collapsed cells. don't know why this is happening. hopefully i can fix it soon. once I solve this bug, I can move on to fixing the issue where patterns are still being applied adjacently to collapsed cells. one thing leads to another. i don't think the current random selection is all that random. it looks like it's picking the last cell it encounters when iterating through the map. so i'm assuming the chaos values aren't right or something. suddenly. i don't know what broke it. but something in this change broke it. i can't think of any fucking thing because the chaos should be recalculated every time you pick the next candidate so why is it like this? fuck this shit. this progress has been so horrible and i get caught up in random bullshit that is not the real problem at hand. how long? HOW FUCKING LONG WILL THIS TAKE? a month. i've been on this generation for a FUCKING month. and it's just random shit after random shit. am i this bad of a programmer? should I have just worked at mcdonald's? i want to break something, this is so infuriating. the more i think about the state of everything, the more i want to break something. there's a glass on my desk right now. it would be so satisfying. it's just so bad. i am genuinely unsure as to why the fuck the candidatePosition broke when i'm working on SelectPattern? idk. just debug through it sometime later. i'll stumble upon the correct answer at some point, right?